refactor: tools.py split Phase 3 — extract MCP/skills to _mcp_skills.py
Moved MCP server registration and prompt-only skill management tools
(~345 lines) from _core.py into _mcp_skills.py.  Self-contained — only
depends on paths + stdlib (json, os, re, shutil).

_core.py now ~4050 lines (down from ~5269 originally).
